streamline title flow (route.meta for list, explicit for details)
