+++ src/custom/culture/views/DepsView.vue
@@ -8,7 +8,6 @@ import { fromMarkdown } from '@/utils'
 import { useMeta } from '@/utils/seo'
 
 useMeta({
-  title: () => undefined,
   description: () => config.website.homepage.meta_description,
   canonicalUrl: () => `${window.location.origin}/deps`
 })
