+++ src/utils/seo.ts
@@ -1,11 +1,16 @@
 import { useHead } from '@unhead/vue'
 import type { Ref } from 'vue'
-import { toValue } from 'vue'
+import { inject, toValue, watch } from 'vue'
 import type { RouteLocationRaw } from 'vue-router'
-import { useRouter } from 'vue-router'
+import { useRoute, useRouter } from 'vue-router'
 
 import config from '@/config'
+import {
+  AccessibilityPropertiesKey,
+  type AccessibilityPropertiesType
+} from '@/model/injectionKeys'
 import { stripFromMarkdown } from '@/utils'
+import { DebugLogger } from '@/utils/debug'
 
 type ReactiveInput<T> = Ref<T> | (() => T)
 
@@ -43,15 +48,42 @@ export function useMeta({
   canonicalUrl,
   noIndex
 }: {
-  title: ReactiveInput<string | undefined>
+  title?: ReactiveInput<string | undefined>
   description: ReactiveInput<string | undefined>
   keywords?: ReactiveInput<string[] | undefined>
   canonicalUrl: ReactiveInput<string | undefined | null>
   noIndex?: ReactiveInput<boolean | undefined>
 }) {
+  const route = useRoute()
+  const setAccessibilityProperties = inject(
+    AccessibilityPropertiesKey
+  ) as AccessibilityPropertiesType
+
+  if (title !== undefined && route.meta.title) {
+    new DebugLogger('useMeta').warn(
+      `Both route.meta.title ("${route.meta.title}") and explicit title are set on "${route.path}". ` +
+        'This causes a mismatch between the live region announcement and document.title. Remove one.'
+    )
+  }
+
+  // Explicit title takes precedence; fall back to route.meta.title for pages
+  // that define their title in the router (list pages, static pages, home)
+  const resolvedTitle = () => toValue(title) ?? (route.meta.title || undefined)
+
   const fullTitle = () => {
-    const t = toValue(title)
+    const t = resolvedTitle()
     return t ? `${t} | ${config.website.title}` : config.website.title
+  }
+
+  // Only announce when App.vue's route watcher won't (routes without a static meta.title)
+  if (!route.meta.title) {
+    watch(
+      resolvedTitle,
+      (t) => {
+        if (t) setAccessibilityProperties(t)
+      },
+      { immediate: true }
+    )
   }
 
   const args = {
