+++ src/views/dataservices/DataserviceDetailView.vue
@@ -14,10 +14,6 @@ import GenericContainer from '@/components/GenericContainer.vue'
 import MetricsStatBoxes from '@/components/MetricsStatBoxes.vue'
 import OrganizationLogo from '@/components/OrganizationLogo.vue'
 import VIconCustom from '@/components/VIconCustom.vue'
-import {
-  AccessibilityPropertiesKey,
-  type AccessibilityPropertiesType
-} from '@/model/injectionKeys'
 import { useCurrentPageConf, useRouteParamsAsString } from '@/router/utils'
 import { useDataserviceStore } from '@/store/DataserviceStore'
 import { descriptionFromMarkdown, formatDate } from '@/utils'
@@ -65,10 +61,6 @@ const goToPage = (page: number) => {
   currentPage.value = page + 1
   loadDatasets()
 }
-
-const setAccessibilityProperties = inject(
-  AccessibilityPropertiesKey
-) as AccessibilityPropertiesType
 
 const links = computed(() => {
   const breadcrumbs = [{ to: '/', text: 'Accueil' }]
@@ -122,14 +114,6 @@ useMeta({
   description: () => dataservice.value?.description,
   canonicalUrl: () => dataservice.value?.self_web_url
 })
-
-watch(
-  pageTitle,
-  (t) => {
-    if (t) setAccessibilityProperties(t)
-  },
-  { immediate: true }
-)
 
 onMounted(() => {
   dataserviceStore
